Flatten duplicate JWT exp parsing into a single ./next helper
next/VsCodeExtensionCredential had a private getTokenExpiry that duplicated the legacy
utils/tryGetTokenExpiration JWT exp parsing. Extract the canonical implementation to
next/utils/tryGetTokenExpiration.ts (getTokenExpiry -> { expiresOnTimestamp, refreshAfterTimestamp }).
The credential now imports it, and the legacy tryGetTokenExpiration(session) becomes a thin wrapper
returning .expiresOnTimestamp.
